Feature: Move viewsets to views.py
